Support for publishing to GPR
This commit addresses issue #2, by adding support for publishing the gem
to the GitHub Package Registry (GPR).

In order to publish to GPR, Github requires the gem name and the repo
name to be the same. To avoid disrupting the current deployment, this is
a fork of the old `ds-api-ruby`, leaving that gem in place for legacy
support.