Improve logging of API calls
epimorphics/hmlr-linked-data#85

This commit removes the use of automated Faraday logging of remote API
calls. The issue is that Faraday does not log the API requests in a way
that matches our local best practice.

Instead, we manually log the API call **after** it has completed. That
way, we can associate the HTTP URL with the return status and call
duration, in one single log line.